requirements.txt
opencv-python
dlib
scipy.spatial
numpy>=1.23.0
imutils
pygame
threading